refactor: Create base module to define base schemas
