+++ api/schema/default.py
@@ -1,24 +1,6 @@
 from marshmallow import Schema, fields
 
 
-class SymbolRequiredSchema(Schema):
-    symbol = fields.String(
-        required=True,
-    )
-
-
-class SymbolOptionalSchema(Schema):
-    symbol = fields.String(
-        required=False,
-    )
-
-
-class PairRequiredSchema(Schema):
-    symbol = fields.String(
-        required=True,
-    )
-
-
 class TimeSchema(Schema):
     serverTime = fields.Number(
         required=True,
